Finished 1

+++ kernel/defs.h
@@ -108,8 +108,8 @@ int             either_copyin(void *dst, int user_src, uint64 src, uint64 len);
 void            procdump(void);
 
 struct proc*    proc_by_pid(int);
-int             map_shared_pages(struct proc*, struct proc*, uint64, uint64);
-int             unmap_shared_pages(struct proc*, uint64, uint64);
+uint64             map_shared_pages(struct proc*, struct proc*, uint64, uint64);
+uint64             unmap_shared_pages(struct proc*, uint64, uint64);
 
 // swtch.S
 void            swtch(struct context*, struct context*);
